[Private] Our group doesn't broadcast if there is a command
Playing with usability.

The chat will broadcast to the user if there is only one
If a second user joins, only reply to the user will lock the chat to
that user

If someone in the group calls a command, we go into "private mode"
So the users don't get the messages.

A reply will get the system out of the private mode.

Reply to users in a group is a bit of a pain in telegram.
Since you have to select the user everytime you want to reply to it
(makes sense)

Just playing with this concept, we might just scrap everything and go
with the default behaviour, hence not caring much about the code :D

Signed-off-by: Sergio A. Mart <sergio@engineer.blue>